Store user_ids as BIGINT everywhere

+++ init_db.py
@@ -10,9 +10,9 @@ try:
     CON.execute('''CREATE TABLE chat (
                     chat_id BIGINT PRIMARY KEY,
                     title VARCHAR(128),
-                    admins INTEGER[],
-                    prev_winners INTEGER[],
-                    cur_winner INTEGER
+                    admins BIGINT[],
+                    prev_winners BIGINT[],
+                    cur_winner BIGINT
                   )''')
 
     CON.execute('''CREATE TABLE chat_user (
